Cog refactor
