+++ donguri_gaeru/cogs/player.py
@@ -16,14 +16,106 @@
 
 import logging
 
+import discord
 from bot import DonguriGaeruBot
-from database import Player
+from database import Match, Player
 from discord.ext import commands
 from sqlalchemy import select, update
 from sqlalchemy.exc import IntegrityError, NoResultFound
 from utils import helpers
 
 from donguri_gaeru import Session
+
+
+class PlayerMatchCog(commands.Cog):
+    """
+    Commands for submitting, fixing, and deleting matches from the database.
+    """
+
+    def __init__(self, bot: DonguriGaeruBot):
+        self.bot = bot
+        self.log = logging.getLogger("donguri_gaeru")
+
+    @commands.command()
+    async def submit(
+        self,
+        ctx: commands.Context,
+        score1: int,
+        score2: int,
+        player2: helpers.NameOrUser,
+    ):
+        with Session() as session:
+            stmt = select(Player)
+            playerA = (
+                session.execute(stmt.where(Player.discord == ctx.author.id))
+                .scalars()
+                .first()
+            )
+
+            if playerA is None:
+                await ctx.send(
+                    "You're not registered! "
+                    f"Run `{ctx.prefix}register <name>` to register."
+                )
+                return
+
+            if isinstance(player2, str):
+                playerB: Player = (
+                    session.execute(stmt.where(Player.name.ilike(player2)))
+                    .scalars()
+                    .first()
+                )
+
+                if playerB is None:
+                    playerB = Player(name=player2)
+                    session.add(playerB)
+
+            else:
+                playerB: Player = (
+                    session.execute(stmt.where(Player.discord == player2.id))
+                    .scalars()
+                    .first()
+                )
+
+                if playerB is None:
+                    await ctx.send(
+                        "Player 2 isn't registered! "
+                        f"Run `{ctx.prefix}register <name>` to register."
+                    )
+                    return
+
+            msg: discord.Message = await ctx.send(
+                "Is this information correct?\n\n"
+                f"({playerA.name}) {score1} - {score2} ({playerB.name})"
+            )
+
+            if await helpers.confirmation(ctx, msg):
+                session.commit()
+
+                session.refresh(playerB)
+
+                match = Match(
+                    playerA_id=playerA.id,
+                    playerB_id=playerB.id,
+                    scoreA=score1,
+                    scoreB=score2,
+                )
+                session.add(match)
+                session.commit()
+
+                session.refresh(match)
+
+                match_string = (
+                    f"`{match.id}`: "
+                    f"({match.playerA.name}) {match.scoreA} - "
+                    f"{match.scoreB} ({match.playerB.name})\n"
+                    f"on {match.created}"
+                )
+                usr = f"{ctx.author.id} ({ctx.author.name}#{ctx.author.discriminator})"
+                self.log.info(f"{usr} submitted match:\n{match_string}")
+                await ctx.send(f"Submitted match!\n\n{match_string}")
+            else:
+                session.rollback()
 
 
 class RegistrationCog(commands.Cog):
@@ -88,6 +180,8 @@ class PlayerSettingsCog(commands.Cog):
                 await ctx.send("That's already your name!")
                 return
 
+            old_name = player.name
+
             stmt = update(Player).where(Player.discord == ctx.author.id)
 
             try:
@@ -99,10 +193,11 @@ class PlayerSettingsCog(commands.Cog):
             else:
                 session.commit()
 
-            self.log.info(f"{player.name} changed their name to {name}!")
+            self.log.info(f"{old_name} changed their name to {name}!")
             await ctx.send(f"Your name was changed to {name}! Hello, {name}!")
 
 
 def setup(bot: DonguriGaeruBot):
+    bot.add_cog(PlayerMatchCog(bot))
     bot.add_cog(RegistrationCog(bot))
     bot.add_cog(PlayerSettingsCog(bot))
